be/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
google-genai>=0.5
pydantic>=2.9,<3.0
pydantic-settings>=2.6,<3.0

# Auth & DB (stage 2)
sqlalchemy>=2.0,<3.0
python-jose[cryptography]>=3.3,<4.0
bcrypt>=4.0,<5.0
email-validator>=2.1,<3.0
